Refactor database access to use service layer and add logging functionality
