+++ package.json
@@ -4,7 +4,7 @@
 	"version": "0.0.1",
 	"type": "module",
 	"scripts": {
-		"dev": "vite dev",
+		"dev": "vite dev | pino-pretty",
 		"build": "vite build",
 		"preview": "vite preview",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
@@ -62,6 +62,8 @@
 		"@typescript-eslint/parser": "^8.18.2",
 		"drizzle-orm": "^0.38.3",
 		"drizzle-typebox": "^0.2.1",
-		"eslint-plugin-drizzle": "^0.2.3"
+		"eslint-plugin-drizzle": "^0.2.3",
+		"pino": "^9.6.0",
+		"pino-pretty": "^13.0.0"
 	}
 }
