+++ src/lib/server/logger.ts
@@ -1,0 +1,16 @@
+import pino from 'pino';
+
+const logger = pino({
+	level: process.env.NODE_ENV === 'production' ? 'info' : 'debug',
+	transport:
+		process.env.NODE_ENV !== 'production'
+			? {
+					target: 'pino-pretty',
+					options: {
+						colorize: true
+					}
+				}
+			: undefined
+});
+
+export { logger };
