+++ src/routes/accounts/create/+page.server.ts
@@ -1,4 +1,4 @@
-import { createAccount } from '$lib/server/db/accounts';
+import { createAccount } from '$lib/server/services/accountsService';
 import { insertAccountSchema, type InsertAccount } from '$lib/server/db/schema/accounts';
 import { fail, redirect } from '@sveltejs/kit';
 import { superValidate } from 'sveltekit-superforms';
