+++ src/routes/categories/+page.server.ts
@@ -1,4 +1,8 @@
-import { createCategory, deleteCategory, getCategories } from '$lib/server/db/categories';
+import {
+	createCategory,
+	deleteCategory,
+	getCategories
+} from '$lib/server/services/categoriesService';
 import type { InsertCategory } from '$lib/server/db/schema/categories';
 import type { Actions, PageServerLoad } from './$types';
